Harden the advertisement code
- remove unwraps from `send_advert()` function body
- handle interrupts when trying to write advert to socket
- log appropriately when advert fails

There is still the outstanding issue of doing some tracking of
advertisement failures. The original UCarp code is doing this, but it is
not clear from reading the code what the point is. There is some
handling of a flag `carp_suppress_preempt` and there is a a call to
`carp_send_ad_all()` when some max error count is hit. The
`carp_send_ad_all` function does not do anything except call
`carp_send_ad()` again though. This all looks incomplete.

Failed advertisements should be handled by the remote backup node's
timeout logic. The only danger I can see is that the existing primary
starts failing to send advertisements, but continues to think it is the
primary node. This would mean two nodes are primary. The node that
becomes primary last will issue the arp and try to take over the shared
VIP. However, if we have a network partition then the arp might never
make it to the entire network. There is some more to think about when it
comes to handling advertisement errors.